chore: save logs in SettingsStore
